+++ src/applications/vre/28-1900/orientation/OrientationApp.jsx
@@ -1,44 +1,85 @@
 import React, { useState, useEffect } from 'react';
+import Scroll from 'react-scroll';
+import { focusElement } from 'platform/utilities/ui';
 import StepComponent from './StepComponent';
 import { orientationSteps } from './utils';
+import {
+  CHAPTER_31_ROOT_URL,
+  WIZARD_STATUS_COMPLETE,
+} from 'applications/vre/28-1900/constants';
+
+const scroller = Scroll.scroller;
+const scrollToTop = () => {
+  scroller.scrollTo('topScrollElement', {
+    duration: 500,
+    delay: 0,
+    smooth: true,
+  });
+};
 
 const OrientationApp = props => {
   const [step, setStep] = useState(0);
-  const { formStartHandler, formControlStatus } = props;
+  const [formStartControl, setFormStartControl] = useState(false);
+  const { wizardStateHandler } = props;
 
-  // if a user has gone through the orientation already, reset this so the form start controls are once again hidden.
+  // do this to prevent focus skip on prior link tag
   useEffect(() => {
-    if (formControlStatus) {
-      formStartHandler(false);
+    if (formStartControl) {
+      focusElement('#FormStartControl');
+    } else if (step > 0) {
+      focusElement('#StepTitle');
+      scrollToTop();
     }
-  }, []);
+  });
+
   return (
-    <div className="row vads-u-margin-bottom--1 vads-u-margin-top--2 vads-u-border--1px vads-u-padding--3 orientation-border">
-      <StepComponent step={step} />
-      <div>
-        {step > 0 && (
-          <a
-            onClick={() => (step === 0 ? step : setStep(step - 1))}
-            className="usa-button usa-button-secondary vads-u-padding-x--4"
+    <>
+      <div className="row vads-u-margin-bottom--1 vads-u-margin-top--2 vads-u-border--1px vads-u-padding--3 orientation-border">
+        <StepComponent step={step} />
+        <div>
+          {step > 0 && (
+            <button
+              onClick={() => (step === 0 ? step : setStep(step - 1))}
+              className="usa-button usa-button-secondary vads-u-padding-x--4"
+            >
+              « Back
+            </button>
+          )}
+          <button
+            onClick={() => {
+              if (step < orientationSteps.length - 1) {
+                setStep(step + 1);
+              } else {
+                setFormStartControl(true);
+              }
+            }}
+            type="button"
+            className="usa-button-primary"
           >
-            « Back
-          </a>
-        )}
-        <a
-          onClick={() => {
-            if (step < orientationSteps.length - 1) {
-              setStep(step + 1);
-            } else {
-              formStartHandler(true);
-            }
-          }}
-          type="button"
-          className="usa-button-primary"
-        >
-          {step === 0 ? 'Start VR&E orientation' : 'Continue »'}
-        </a>
+            {step === 0 ? 'Start VR&E orientation' : 'Continue »'}
+          </button>
+        </div>
       </div>
-    </div>
+      {formStartControl && (
+        <div className="vads-u-padding--3 vads-u-background-color--gray-lightest">
+          <p>
+            <strong>Thank you for viewing the VR&E orientation.</strong> To
+            apply for Veteran Readiness & Employment benefits now, click the
+            button below.
+          </p>
+          <a
+            id="FormStartControl"
+            href={CHAPTER_31_ROOT_URL}
+            className="usa-button-primary va-button-primary"
+            onClick={() => {
+              wizardStateHandler(WIZARD_STATUS_COMPLETE);
+            }}
+          >
+            Apply for Veteran Readiness and Employment
+          </a>
+        </div>
+      )}
+    </>
   );
 };
